+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import SearchPage from "./page/SearchPage";
+import SearchPage from "./pages/SearchPage";
 
 function App() {
   return (
